Use tab type (vs. tab name) to identify current tab for highlighting
A tab's name is not necessarily unique, so to make sure we don't end up
with multiple highlighted tabs, use the tab alist's CAR (either 'tab or
'current-tab) to identify the current tab.

This assumes that the result of calling the function referenced by
`tab-bar-tabs-function` has the same structure as the one that's
referenced by default, `tab-bar-tabs`.